move locator filters implemenation to their own files
