+++ src/app/qgisapp.cpp
@@ -260,7 +260,16 @@ Q_GUI_EXPORT extern int qt_defaultDpiX();
 #include "qgslayoutguiutils.h"
 #include "qgslocatorwidget.h"
 #include "qgslocator.h"
-#include "qgsinbuiltlocatorfilters.h"
+#include "qgsactionlocatorfilter.h"
+#include "qgsactivelayerfeatureslocatorfilter.h"
+#include "qgsalllayersfeatureslocatorfilter.h"
+#include "qgsbookmarklocatorfilter.h"
+#include "qgsexpressioncalculatorlocatorfilter.h"
+#include "qgsgotolocatorfilter.h"
+#include "qgslayertreelocatorfilter.h"
+#include "qgslayoutlocatorfilter.h"
+#include "qgsnominatimlocatorfilter.h"
+#include "qgssettingslocatorfilter.h"
 #include "qgslogger.h"
 #include "qgsmapcanvas.h"
 #include "qgsmapcanvasdockwidget.h"
